requirements.txt
python-binance==1.0.17
ccxt==4.0.89
pandas==1.5.3
numpy==1.25.2
python-dotenv==1.0.0
pytest==7.4.2
requests==2.31.0
